Add helper class to remove duplicate code
