+++ lib/perf-top/metrics/util/generate-data.js
@@ -13,10 +13,7 @@
  * permissions and limitations under the License.
  */
 
-var http = require('http');
-var https = require('https');
-var url = require('url');
-
+var helper = require("../../helper");
 /**
  * Makes a HTTP(S) request to "{endpoint}/_opendistro/_performanceanalyzer/metrics?metrics=${metrics}&agg=${aggregates}&dim=${dimensions}&nodes=all"
  * and parses the response to a hashmap object.
@@ -32,62 +29,15 @@ function getMetricData (endpoint, metrics, aggregates, dimensions, done) {
   var aggParam = (aggregates) ? `&agg=${aggregates}` : '&agg=';
   var dimParam = (dimensions) ? `&dim=${dimensions}` : '&dim=';
 
-  var urlOptions = getURLOptions(endpoint, `/_opendistro/_performanceanalyzer/metrics?${metricParam}${aggParam}${dimParam}&nodes=all`);
+  var urlOptions = helper.getURLOptions(endpoint, `/_opendistro/_performanceanalyzer/metrics?${metricParam}${aggParam}${dimParam}&nodes=all`);
 
-  makeRequest(urlOptions, function (response) {
+  helper.makeRequest(urlOptions, function (response) {
     if (response === '') {
       done({});
     } else {
       done(getDataPerNode(response));
     }
   });
-}
-
-/**
- * Gets a URL option with host, port (default = 80), and path for http(s).request().
- *
- * @param {string} endpoint - endpoint for the query.
- * @param {string} path - metric query request path.
- */
-function getURLOptions (endpoint, path) {
-  endpoint = ( endpoint.startsWith('http://') || endpoint.startsWith('https://') ) ? endpoint : 'http://' + endpoint;
-  var parsedURL = url.parse(url.resolve(endpoint, path));
-  return {
-    protocol: parsedURL.protocol,
-    host: parsedURL.hostname,
-    auth: parsedURL.auth,
-    port: (parsedURL.port === null) ? (parsedURL.protocol === 'https:' ? 443 : 80 ): parsedURL.port,
-    path: parsedURL.path
-  };
-}
-
-/**
- * Makes a HTTP(S) request and returns the response.
- *
- * @param {string} urlOptions - URL option for the HTTP(S) request.
- * @param {string} done - callback for the HTTP(S) response.
- */
-function makeRequest (urlOptions, done) {
-  var rawData = '';
-  var respond = function (response) {
-    response.on('data', function (chunk) {
-      rawData += chunk;
-    });
-    response.on('end', function () {
-      done(rawData);
-    });
-  };
-  var client = http;
-  if (urlOptions.protocol == "https:") {
-    client = https;
-    urlOptions.rejectUnauthorized = false;
-  }
-  var req = client.request(urlOptions, respond);
-  req.on('error', function (error) {
-    console.error(error);
-    done('');
-  });
-  req.end();
 }
 
 /**
@@ -98,8 +48,8 @@ function makeRequest (urlOptions, done) {
  * @param {string} done - callback for the HTTP(S) response.
  */
 function getMetricUnits (endpoint, done) {
-  var urlOptions = getURLOptions(endpoint, '/_opendistro/_performanceanalyzer/metrics/units');
-  makeRequest(urlOptions, function (response) {
+  var urlOptions = helper.getURLOptions(endpoint, '/_opendistro/_performanceanalyzer/metrics/units');
+  helper.makeRequest(urlOptions, function (response) {
     if (response === '') {
       console.error('Failed to retrieve units for metrics. HTTP(S) response was empty.');
       done({});
@@ -118,30 +68,6 @@ function getMetricUnits (endpoint, done) {
       done({});
     }
   });
-}
-
-/**
- * Convert and parse a string to a number.
- *
- * @param {string} data - string
- * @returns {number|string} - if `data` cannot be converted to a number, return `data` back,
- *                            else return the number (with maximum of 2 decimal points if less than 100,
- *                            else a rounded integer).
- */
-function parseNumberData (data) {
-  if (data === null) {
-    return 'null';
-  } else if (data === '') {
-    return data;
-  }
-  var parsedData = Math.round(data * 100) / 100;
-  if (isNaN(parsedData)) {
-    return data;
-  } else if (parsedData < 100) {
-    return parsedData;
-  } else {
-    return Math.round(parsedData);
-  }
 }
 
 /**
@@ -263,7 +189,7 @@ function getDataPerNode (rawData) {
       nodeDimensions.push(field.name);
     });
     jsonData[nodeName].data.records.forEach(function (record) {
-      nodeData.push(record.map(parseNumberData));
+      nodeData.push(record.map(helper.parseNumberData));
     });
 
     allData[nodeName] = { fields: nodeDimensions, data: nodeData, timestamp: jsonData[nodeName].timestamp };
@@ -375,7 +301,6 @@ module.exports.getMetricUnits = getMetricUnits;
 module.exports.getNodeData = getNodeData;
 module.exports.getDimensionData = getDimensionData;
 module.exports.aggregateMetricData = aggregateMetricData;
-module.exports.parseNumberData = parseNumberData;
 module.exports.removeStaleData = removeStaleData;
 module.exports.sortDataByDecreasingOrder = sortDataByDecreasingOrder;
 module.exports.addCommaDelimiter = addCommaDelimiter;
