+++ lib/perf-top/rca/util/temperature-profile/generate-data.js
@@ -13,9 +13,7 @@
  * permissions and limitations under the License.
  */
 
-var http = require('http');
-var https = require('https');
-var url = require('url');
+var helper = require('../../../helper.js');
 
 /**
  * Makes a HTTP(S) request to "{endpoint}/_opendistro/_performanceanalyzer/rca?${nameParam}${localParam}"
@@ -32,12 +30,12 @@ var url = require('url');
 function getMetricData(endpoint, name, local, dimension, graphParams, type, done) {
   var nameParam = name ? `name=${name}` : "name=AllTemperatureDimensions";
   var localParam = local ? `&local=${local}` : "&local=true";
-  var urlOptions = getURLOptions(
+  var urlOptions = helper.getURLOptions(
     endpoint,
     `/_opendistro/_performanceanalyzer/rca?${nameParam}${localParam}`
   );
 
-  makeRequest(urlOptions, function (response) {
+  helper.makeRequest(urlOptions, function (response) {
     if (response === "") {
       console.error("empty response from server");
       done({});
@@ -45,78 +43,6 @@ function getMetricData(endpoint, name, local, dimension, graphParams, type, done
       done(getData(response, dimension, graphParams, type));
     }
   });
-}
-
-/**
- * Gets a URL option with host, port (default = 80), and path for http(s).request().
- *
- * @param {string} endpoint - endpoint for the query.
- * @param {string} path - metric query request path.
- */
-function getURLOptions (endpoint, path) {
-  endpoint = ( endpoint.startsWith('http://') || endpoint.startsWith('https://') ) ? endpoint : 'http://' + endpoint;
-  var parsedURL = url.parse(url.resolve(endpoint, path));
-  return {
-    protocol: parsedURL.protocol,
-    host: parsedURL.hostname,
-    auth: parsedURL.auth,
-    port: (parsedURL.port === null) ? (parsedURL.protocol === 'https:' ? 443 : 80 ): parsedURL.port,
-    path: parsedURL.path
-  };
-}
-
-/**
- * Makes a HTTP(S) request and returns the response.
- *
- * @param {string} urlOptions - URL option for the HTTP(S) request.
- * @param {string} done - callback for the HTTP(S) response.
- */
-function makeRequest (urlOptions, done) {
-  var rawData = '';
-  var respond = function (response) {
-    response.on('data', function (chunk) {
-      rawData += chunk;
-    });
-    response.on('end', function () {
-      done(rawData);
-    });
-  };
-  var client = http;
-  if (urlOptions.protocol == "https:") {
-    client = https;
-    urlOptions.rejectUnauthorized = false;
-  }
-  var req = client.request(urlOptions, respond);
-  
-  req.on('error', function (error) {
-    console.error(error);
-    done('');
-  });
-  req.end();
-}
-
-/**
- * Convert and parse a string to a number.
- *
- * @param {string} data - string
- * @returns {number|string} - if `data` cannot be converted to a number, return `data` back,
- *                            else return the number (with maximum of 2 decimal points if less than 100,
- *                            else a rounded integer).
- */
-function parseNumberData (data) {
-  if (data === null) {
-    return 'null';
-  } else if (data === '') {
-    return data;
-  }
-  var parsedData = Math.round(data * 100) / 100;
-  if (isNaN(parsedData)) {
-    return data;
-  } else if (parsedData < 100) {
-    return parsedData;
-  } else {
-    return Math.round(parsedData);
-  }
 }
 
 /**
@@ -170,7 +96,7 @@ function getData(rawData, dimension, graphParams, type) {
       var field = node.dimension + " - " + graphParams;
       nodeDimensions.push(field);
       if (node.hasOwnProperty(graphParams)) {
-        nodeData.push(parseNumberData(node[graphParams]));
+        nodeData.push(helper.parseNumberData(node[graphParams]));
       } else {
         nodeData.push("null");
       }
@@ -190,7 +116,7 @@ function getData(rawData, dimension, graphParams, type) {
               tmpData.push(tmp.index_name);
               tmpData.push(tmp.shard_id);
               tmp.temperature.forEach(function (data) {
-                tmpData.push(parseNumberData(data.value));
+                tmpData.push(helper.parseNumberData(data.value));
               });
               nodeData.push(tmpData);
             }
@@ -201,7 +127,7 @@ function getData(rawData, dimension, graphParams, type) {
     } else if (type == "donuts") {
       for (var j = 0; j < zoneSummary.length; j++) {
         nodeDimensions.push(zoneSummary[j].zone);
-        nodeData.push(parseNumberData(zoneSummary[j].all_shards.length));
+        nodeData.push(helper.parseNumberData(zoneSummary[j].all_shards.length));
       }
     }
     break;
@@ -257,6 +183,5 @@ function removeStaleData(metricData, metricGraph) {
   */
 }
 
-module.exports.parseNumberData = parseNumberData;
 module.exports.getMetricData = getMetricData;
 module.exports.removeStaleData = removeStaleData;
